refactor: avoid unnecessary re-renders and prop drilling
- Keeping 2 components in same file for clarity and readability
